Add strict, native op + test & refactor
